+++ probdiffeq/_adaptive.py
@@ -4,8 +4,9 @@ from typing import Any, Generic, TypeVar
 import jax
 import jax.numpy as jnp
 
-from probdiffeq import controls, ivpsolvers
+from probdiffeq import controls
 from probdiffeq.backend import containers
+from probdiffeq.ivpsolvers import solver
 
 
 class _RejectionState(containers.NamedTuple):
@@ -25,7 +26,7 @@ def _reference_state_fn_max_abs(sol, sol_previous):
     return jnp.maximum(jnp.abs(sol), jnp.abs(sol_previous))
 
 
-T = TypeVar("T", bound=ivpsolvers.Solver)
+T = TypeVar("T", bound=solver.Solver)
 """A type-variable for (non-adaptive) IVP solvers."""